:construction: CRUD clases, no muestra listado de maestros

+++ app/routes/alumnos_router.py
@@ -1,5 +1,5 @@
 
-from flask import render_template, redirect, url_for, flash, request
+from flask import render_template, redirect, session, url_for, flash, request
 from app import db
 from app.forms.alumno_form import AlumnoForm
 from app.models.alumno_model import Alumno
@@ -8,12 +8,18 @@ def configurar_alumnos(app):
     # Ruta para listar alumnos
     @app.route('/alumnos', methods=['GET'])
     def listar_alumnos():
+        if 'user' not in session:
+            flash('Debes iniciar sesión para acceder al dashboard.', 'warning')
+            return redirect(url_for('login'))
         alumnos = Alumno.query.all()
         return render_template('alumnos/listar.html', alumnos=alumnos)
 
     # Ruta para crear un nuevo alumno
     @app.route('/alumnos/crear', methods=['GET', 'POST'])
     def crear_alumno():
+        if 'user' not in session:
+            flash('Debes iniciar sesión para acceder al dashboard.', 'warning')
+            return redirect(url_for('login'))
         form = AlumnoForm()
         if form.validate_on_submit():
             nuevo_alumno = Alumno(nombre=form.nombre.data, apellido=form.apellido.data, email=form.email.data, grado_id=form.grado.data)
@@ -26,6 +32,9 @@ def configurar_alumnos(app):
     # Ruta para editar un alumno existente
     @app.route('/alumnos/editar/<int:id>', methods=['GET', 'POST'])
     def editar_alumno(id):
+        if 'user' not in session:
+            flash('Debes iniciar sesión para acceder al dashboard.', 'warning')
+            return redirect(url_for('login'))
         alumno = Alumno.query.get_or_404(id)
         form = AlumnoForm(obj=alumno)
         
@@ -43,6 +52,9 @@ def configurar_alumnos(app):
     # Ruta para eliminar un alumno
     @app.route('/alumnos/eliminar/<int:id>', methods=['POST'])
     def eliminar_alumno(id):
+        if 'user' not in session:
+            flash('Debes iniciar sesión para acceder al dashboard.', 'warning')
+            return redirect(url_for('login'))
         alumno = Alumno.query.get_or_404(id)
         db.session.delete(alumno)
         db.session.commit()
